+++ examples/layers.rs
@@ -30,7 +30,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
             grid_size,
             size: tilemap_size,
             storage: tile_storage,
-            texture: TilemapTexture(texture_handle.clone()),
+            texture: TilemapTexture::Single(texture_handle.clone()),
             tile_size,
             transform: get_tilemap_center_transform(&tilemap_size, &grid_size, 0.0),
             ..Default::default()
@@ -54,7 +54,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
             grid_size: grid_size,
             size: tilemap_size,
             storage: tile_storage,
-            texture: TilemapTexture(texture_handle),
+            texture: TilemapTexture::Single(texture_handle),
             tile_size: TilemapTileSize { x: 16.0, y: 16.0 },
             transform: get_tilemap_center_transform(&tilemap_size, &grid_size, 1.0)
                 * Transform::from_xyz(32.0, 32.0, 0.0),
